+++ main_project/component/component.hpp
@@ -19,10 +19,10 @@ public:
     Component(string _name);
     virtual ~Component() = 0; //destructor of base class should be virtual (purely virtual as component should be abstract)
 
-    virtual float getConductance() const;
+    virtual double getConductance() const;
     virtual float getVoltage() const;
-    virtual float getCurrent() const; //For complex components, this returns the current through the companion model's current source rather than through the whole component
-    virtual float getTotalCurrent(float voltage, int order = 1); //For complex components, this return the current through the whole component
+    virtual double getCurrent() const; //For complex components, this returns the current through the companion model's current source rather than through the whole component
+    virtual double getTotalCurrent(float voltage, int order = 1); //For complex components, this return the current through the whole component
 
     string getName() const;
     virtual vector<int> getNodes() const = 0;
