+++ main_project/component/diode.cpp
@@ -38,15 +38,15 @@ Diode::Diode(string _name,float c, int n1, int n2, float timeStep, int order)
 	types.push_back(componentType::nonLinear);
 }
 
-float Diode::getConductance() const{
+double Diode::getConductance() const{
 	return compConductance;
 }
 
-float Diode::getCurrent() const{
+double Diode::getCurrent() const{
 	return -compCurrent; //direction of current source is opposite to direction of diode
 }
 
-float Diode::getTotalCurrent(float voltage, int order){
+double Diode::getTotalCurrent(float voltage, int order){
 	return NAN; //not yet implemented
 }
