+++ main_project/output/nonlinearAnalysis.cpp
@@ -30,7 +30,7 @@ string runNonlinearTransience(Circuit& c, float t){
 
     ///Newton-Raphson loop:
 
-    VectorXf x = VectorXf::Zero(highestNodeNumber + voltageSources.size());
+    VectorXd x = VectorXd::Zero(highestNodeNumber + voltageSources.size());
 
     //variables that are used multiple times in this function
     vector<int> nodes{};
